requirements.txt
pandas
numpy
joblib
sklearn
kafka-python
mlflow
kfp
